move slack agent
