+++ packages/server/customer-os-common-module/service/agent_service.go
@@ -3,18 +3,12 @@ package service
 import (
 	"context"
 
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/common"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/data_fields"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/dto"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/enum"
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/tracing"
-	"github.com/opentracing/opentracing-go"
 )
 
 type AgentService interface {
 	VisitorIDAgent(ctx context.Context, eventData *data_fields.WebsiteVisitEvent)
-	SlackAgent(ctx context.Context, event *dto.FlowAgentEvent) error
-	TimelineAgent(ctx context.Context, event *dto.FlowAgentEvent) error
 	ICPAgent(ctx context.Context, event *dto.FlowAgentEvent) error
 }
 
@@ -26,32 +20,4 @@ func NewAgentService(services *Services) AgentService {
 	return &agentService{
 		services: services,
 	}
-}
-
-func (a *agentService) publishAgentResultEvent(
-	ctx context.Context, flowExecutionId, actionExecutionId string, actionExecutionStatus enum.FlowAgentExecutionStatus, errorMessage *string,
-) error {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "AgentService.publishAgentResultEvent")
-	defer span.Finish()
-	tracing.SetDefaultListenerSpanTags(ctx, span)
-
-	resultEvent := dto.FlowAgentExecutionResultEvent{
-		FlowExecutionID:      flowExecutionId,
-		FlowAgentExecutionID: actionExecutionId,
-		Tenant:               common.GetTenantFromContext(ctx),
-		Status:               actionExecutionStatus,
-		ErrorMessage:         errorMessage,
-	}
-
-	a.services.RabbitMQService.PublishFlowAgentEventResult(ctx, resultEvent)
-
-	return nil
-}
-
-func (a *agentService) EmailAgent(ctx context.Context) error {
-	return nil
-}
-
-func (a *agentService) LinkedinAgent(ctx context.Context) error {
-	return nil
 }
